+++ CHANGELOG.md
@@ -6,7 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Name of the component to `SearchAction` to be more generic.
+
+### Fixed
+- `potentialAction.target`.
+- Protocol in SSR.
+- Minify JSON.
 
 ## [0.0.2] - 2019-08-30
-## Added
+### Added
 - Add Component With Google Search Box Script.